Finish credential management
Delete the playbook which sets the root password, credential management takes
care of it.

Comment out incomplete parameters from the example file so that it can work out
of the box.

Use underscores instead of dashes in key names, Ansible does not like dashes.

Do not use a "rpc" user in the example, there is already such user on RHEL.